Update deploy-dev.yaml
- Added 3 new fields to /wallet_overview: https://hellobink.atlassian.net/browse/LOY-2466
is_fully_pll_linked (example and schema)
pll_linked_payment_accounts (example and schema)
total_payment_accounts (example and schema)

- Made 'status' in PLL links an object, and added 3 new fields to this object in /wallet (loyalty_cards and payment_accounts) and /wallet/loyalty_cards/id (loyalty_cards): https://hellobink.atlassian.net/browse/LOY-2448
State (example and schema)
Slug (example and schema)
Description (example and schema)

Also addressed PB's comments:
"is_fully_pll_linked | booleanIndicates whether or not the given Loyalty Card is linked to all Payment Accounts in the user's wallet." - Make when talking about linking we are clear we mean PLL linking. Especially when we are talking to LBG who use linking to mean someting different. So "Loyalty Card is linked ", becomes "Loyalty Card is PLL linked ".
"The number of payment accounts that are PLL linked to the given Loyalty Card."--
"The total number of payment accounts that the user has in their wallet." - Capitalise Payment Accounts please to be consistent.
"pll_links | object (LoyaltyCardPLLLinks)List of the Payment Accounts with an active PLL link to the Loyalty Card." - We need to change this too to reflect that the pll_links object is no longer just showing active links.
"status | objectThe status of the PLL link:activependinginactive" - That list needs to be moved to the state key in the s`status object.
"description | string - If the Loyalty Card or Payment Account is in a state other than active, there will be text to describe the state of the Card." - state of the PLL link not the Card.